Update to mongo

+++ client/src/redux/features/users/usersGetSlice.js
@@ -1,5 +1,4 @@
 import axios from "axios";
-
 import {
   addUsers,
   deleteUsers,
@@ -21,7 +20,9 @@ import {
   setFollow,
   setUnfollow,
   getUserDataGraphs,
+  getPostsByUserToProfile,
 } from "./usersSlice";
+import { getFollowOfThisUser } from "./utilsUsers";
 
 //obtener los users
 export const getUser = () => {
@@ -39,6 +40,7 @@ export const getUser = () => {
 //crear users
 export const createdUser = (user) => {
   return async (dispatch) => {
+    if (user === undefined) return;
     let response = await axios.post("/user", user);
     dispatch(addUsers(response.data));
   };
@@ -48,6 +50,7 @@ export const createdUser = (user) => {
 export const updateUser = (_id, body) => {
   return async (dispatch) => {
     try {
+      if (_id === undefined || body === undefined) return;
       const response = await axios.put(`/users/${_id}`, body);
       if (response) {
         dispatch(updateUsers());
@@ -59,12 +62,14 @@ export const updateUser = (_id, body) => {
   };
 };
 
+//set user genres
 export const setUserGenres = (body) => {
   return async (dispatch) => {
     try {
       const response = await axios.put(`/users/set/genres`, body);
       if (response) {
         dispatch(setGenres(response.data.genres));
+        dispatch(getByFirebaseId(response.data._id));
         dispatch(getUser());
       }
     } catch (error) {
@@ -77,6 +82,8 @@ export const setUserGenres = (body) => {
 export const deleteUser = (_id) => {
   return async (dispatch) => {
     try {
+      if (_id === undefined) return;
+
       await axios.delete(`/users/${_id}`);
       dispatch(deleteUsers());
       dispatch(getUser());
@@ -89,10 +96,38 @@ export const deleteUser = (_id) => {
 export const getUserById = (_id) => {
   return async (dispatch) => {
     try {
+      if (_id === undefined)
+        return console.log(`_id is undefined in getUserById`);
       const response = await axios.get(`/users/${_id}`);
       dispatch(getById(response.data));
     } catch (error) {
       console.log(error);
+    }
+  };
+};
+
+export const getFollowsByUserId = (_id, setter) => {
+  return async (dispatch) => {
+    if (_id === undefined) return;
+
+    try {
+      const { data } = await axios.get(`/follows/${_id}`);
+      dispatch(setFollow(data));
+    } catch (error) {
+      console.log(error);
+    }
+  };
+};
+
+export const getPostsByUser = (_id) => {
+  return async (dispatch) => {
+    if (_id === undefined) return;
+
+    try {
+      const { data } = await axios.get(`/posts/user/${_id}`);
+      dispatch(getPostsByUserToProfile(data));
+    } catch (err) {
+      console.log(err);
     }
   };
 };
@@ -109,6 +144,7 @@ export const cleanUserState = () => {
 
 export const getUserByFirebaseId = (_id) => {
   return async (dispatch) => {
+    if (_id === undefined) return;
     try {
       const { data } = await axios.get(`/users/idGoogle/${_id}`);
       dispatch(getByFirebaseId(data));
@@ -154,6 +190,7 @@ export const getUserLikes = (_id) => {
 
 export const getUserNotification = (_id) => {
   return async (dispatch) => {
+    if (_id === undefined) return;
     try {
       const response = await axios.get(`/notifications/${_id}`);
       await dispatch(getNotifications(response.data));
@@ -201,7 +238,7 @@ export const setUserFollow = (body) => {
     try {
       const response = await axios.post(`/users/follow`, body);
       if (response) {
-        dispatch(setFollow(response.data.FollowerUsers));
+        dispatch(setFollow(response.data));
         dispatch(getUser());
       }
     } catch (error) {
@@ -215,7 +252,7 @@ export const setUserUnfollow = (body) => {
     try {
       const response = await axios.post(`/users/unfollow`, body);
       if (response) {
-        dispatch(setUnfollow(response.data.FollowerUsers));
+        dispatch(setUnfollow(response.data));
         dispatch(getUser());
       }
     } catch (error) {
